Remove styles.module.scss files and refactor card component styles
This commit removes unused styles.module.scss files in ActionIcon and Button components. It also refactors the classnames in the Card component to adhere to the BEM naming convention. Lastly, the ActionIcon import path is cleaned up in the mainNav component and unnecessary file path ".next/types/**/*.ts" is removed from tsconfig.json.