Refactor component names and styles in application
Several changes were made to improve consistency and clarity in component naming. The component 'ListItemWithIcon' was renamed to 'ListItemIcon', and all associated styles were updated. Furthermore, the 'Index' component in the 'logo' directory was renamed to 'Logo'. Lastly, styling changes were made in the Footer, and the file imports in 'mainNav' were rearranged.