Refactor Redux and GraphQL setup for better scalability
The old implementation of GraphQL queries and Redux setup has been refactored. file graphqlBaseQuery and userApi files were deleted and replaced with the new structure in BaseApi, skysApi, and Store files. They now use the enhanced version of redux-toolkit query and graphql-request for improved maintenance and better scalability in the future. Further adjustments were also made in the newsfeed/page and hooks files to comply with the upgraded setup.