Update package references and improve type checking
This commit addresses the incorrect spelling of a @skys-client/queries package reference across several files, correcting it from 'querys' to 'queries'. It also introduces the '@testing-library/react-hooks' dependence into package.json. In addition to this, the type checking for objects in 'hooks.ts' was improved, replacing 'Array.isArray()' with a more accurate 'isObject()' function.